requirements.txt
alembic==1.0.10
amqp==2.5.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
Blueprints==2.3.0.2
celery==4.3.0
Click==7.0
factory-boy==2.12.0
Faker==1.0.7
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
marshmallow-sqlalchemy-referential==1.0.5
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-flask==0.15.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.4
sqlalchemy-mixins==1.1
text-unidecode==1.2
typing==3.6.6
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
